LayoutTests/js/array-sort-sparse-expected.txt

This tests that arrays and array like objects containing holes are sorted correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testSort([,undefined,0,1]) is true
PASS testSort({length:4,1:undefined,2:0,3:1}) is true
PASS successfullyParsed is true

TEST COMPLETE
